login/requirements.txt
Flask==1.0.3
prometheus_client == 0.7.1
mysql-connector

